+++ ui/all-samples/src/components/ButtonToSamples/ButtonToSamples.js
@@ -1,11 +1,11 @@
 import Button from '@enact/ui/Button';
-import {Link} from 'react-router-dom';
+import {Link} from 'react-router';
 
 import css from './ButtonToSamples.module.less';
 
 const ButtonToSamples = () => (
 	<div className={css.buttonContainer}>
-		<Link to="/" className={css.backLink}>
+		<Link to="../" className={css.backLink}>
 			<Button
 				className={css.backButton}
 				style={{zIndex: 1}}